ci(release): Refine Release Please workflow for PRs and labels
The `extra-label` configuration was removed from `release-please/config.json` as the `automerge` label is now explicitly added in the `post-pr` job of the release workflow.

Additionally, the `post-pr` job now retrieves the original commit message, undoes the commit made by `release-please-action`, and then recommits the `CHANGELOG.md` with the original message. This ensures a cleaner and more accurate Git history for release pull requests by preserving the original commit intent.